requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.4
djoser==2.2.2
django-debug-toolbar==4.2.0
celery==5.3.6
redis==5.0.1
django-celery-email==3.0.0
django-cors-headers==4.3.1